chore: 수정

+++ src/main/java/store/buzzbook/front/common/aop/OrderJwtAop.java
@@ -29,7 +29,7 @@ public class OrderJwtAop {
 		if (Objects.nonNull(authorizationHeader)) {
 			Map<String, Object> claims = jwtService.getInfoMapFromJwt(request, response);
 
-			Long userId = (Long)claims.get(JwtService.USER_ID);
+			Long userId = ((Integer)claims.get(JwtService.USER_ID)).longValue();
 			String loginId = (String)claims.get(JwtService.LOGIN_ID);
 			String role = (String)claims.get(JwtService.ROLE);
